+++ packages/components/components/elvis-header/src/react/mobileMenu/mobileMenuStyles.tsx
@@ -96,7 +96,6 @@ export const ImageContainer = styled.div`
   width: 64px;
   height: 64px;
   border-radius: 50%;
-  background-color: ${getThemeColor('background-element-2')};
   margin-bottom: 8px;
   overflow: hidden;
   display: grid;