+++ packages/elvis-colors/src/themes/darkTheme.ts
@@ -61,7 +61,7 @@ export const darkThemeColors = {
   },
   'grey-colors': {
     'grey-70': {
-      color: '#242424',
+      color: '#262626',
       contrastText: '#EDEDED',
     },
     'grey-60': {
@@ -109,8 +109,12 @@ export const darkTheme = {
       contrast: darkThemeColors['grey-colors']['grey-10'].contrastText,
     },
     'text-3': {
-      hex: darkThemeColors['grey-colors']['grey-20'].color,
-      contrast: darkThemeColors['grey-colors']['grey-20'].contrastText,
+      hex: darkThemeColors['primary-colors']['white'].color,
+      contrast: darkThemeColors['primary-colors']['white'].contrastText,
+    },
+    'text-4': {
+      hex: darkThemeColors['primary-colors']['black'].color,
+      contrast: darkThemeColors['primary-colors']['black'].contrastText,
     },
     'text-disabled-1': {
       hex: darkThemeColors['grey-colors']['grey-40'].color,
@@ -119,6 +123,10 @@ export const darkTheme = {
     'text-disabled-2': {
       hex: darkThemeColors['grey-colors']['grey-60'].color,
       contrast: darkThemeColors['grey-colors']['grey-60'].contrastText,
+    },
+    'text-placeholder-1': {
+      hex: darkThemeColors['grey-colors']['grey-20'].color,
+      contrast: darkThemeColors['grey-colors']['grey-20'].contrastText,
     },
   },
   background: {
@@ -142,12 +150,16 @@ export const darkTheme = {
       contrast: darkThemeColors['grey-colors']['grey-60'].contrastText,
     },
     'background-element-4': {
-      hex: darkThemeColors['primary-colors']['grey'].color,
-      contrast: darkThemeColors['primary-colors']['grey'].contrastText,
+      hex: darkThemeColors['grey-colors']['grey-70'].color,
+      contrast: darkThemeColors['grey-colors']['grey-70'].contrastText,
     },
     'background-element-5': {
       hex: darkThemeColors['grey-colors']['grey-70'].color,
       contrast: darkThemeColors['grey-colors']['grey-70'].contrastText,
+    },
+    'background-element-6': {
+      hex: darkThemeColors['primary-colors']['white'].color,
+      contrast: darkThemeColors['primary-colors']['white'].contrastText,
     },
     'background-overlay-1': {
       hex: darkThemeColors['grey-colors']['grey-70'].color,
@@ -313,7 +325,7 @@ export const darkTheme = {
       contrast: darkThemeColors['primary-colors'].white.contrastText,
     },
   },
-  static: {
+  assorted: {
     'static-white': {
       hex: darkThemeColors['primary-colors'].white.color,
       contrast: darkThemeColors['primary-colors'].white.contrastText,
@@ -321,6 +333,13 @@ export const darkTheme = {
     'static-black': {
       hex: darkThemeColors['primary-colors'].black.color,
       contrast: darkThemeColors['primary-colors'].black.contrastText,
+    },
+    'brand-accent': {
+      hex: darkThemeColors['primary-colors'].green.color,
+      contrast: darkThemeColors['primary-colors'].green.contrastText,
+    },
+    'focus-outline': {
+      hex: darkThemeColors['internal-colors']['focus-outline'].color,
     },
   },
 } as const satisfies Theme;