+++ packages/web/src/app/doc-pages/other-pages/color/color-token-table/color-token-table.component.ts
@@ -11,6 +11,7 @@ import {
   signalColors,
   dataColors,
   iconColors,
+  assortedColors,
 } from './colors';
 
 @Component({
@@ -29,5 +30,6 @@ export class ColorTokenTableComponent {
   borderColorsStates = borderColorsStates;
   signalColors = signalColors;
   dataColors = dataColors;
+  assortedColors = assortedColors;
   iconColors = iconColors;
 }